[Refactor](DsOmegaUtils): Separated simple_plugin_setup into atomic utilities.
