+++ lua/ds_omega/utils/apply_plugin_keymappings.lua
@@ -1,0 +1,9 @@
+--- 
+---@param keymappings (table<Mode, table>)
+local function apply_plugin_keymappings(keymappings)
+  for mode, keymappings_for_mode in pairs(keymappings) do
+    require('config.Ui.which_key.utils').apply_keymappings_once_ready(mode, keymappings_for_mode)
+  end
+end
+
+return apply_plugin_keymappings
