+++ lua/ds_omega/utils/init.lua
@@ -1,82 +1,3 @@
-local prequire = require('utils').prequire
-
-local function prequire_plugin(plugin_name)
-  local plugin_is_available, plugin = prequire(plugin_name)
-
-  if not plugin_is_available then
-    notify(
-      'Plugin `' .. plugin_name .. '` does not exist!',
-      vim.log.levels.ERROR,
-      {
-        title = 'Core',
-      }
-    )
-  end
-
-  return plugin_is_available, plugin
-end
-
---- Perform common setup actions for connecting layer with user configuration.
----@param plugin_name 
----@param config_relative_path (string) If only layer name is given, plugin_name will be used
--- instead assuming config is nested in this layer under the same name as
--- plugin_name.
----@example
--- -- Config is in `Integrations.toggleterm`.
--- simple_plugin_setup('toggleterm', 'Integrations')
---
--- -- Take config without changes: `Navigation.other`.
--- simple_plugin_setup('other-nvim', 'Navigation.other')
----@return
-local function simple_plugin_setup(plugin_name, config_relative_path)
-  local plugin_is_available, plugin = prequire_plugin(plugin_name)
-  if not plugin_is_available then
-    return
-  end
-
-  -- Finding first segment. It must be a layer.
-  local segment_end = string.find(config_relative_path, "%.")
-
-  ---  Only layer name was given.
-  -- Assuming plugin_name and config_name for this plugin are the same.
-  if not segment_end then
-    local layer = config_relative_path
-
-    config_relative_path = layer ..'.'.. plugin_name
-  end
-
-  local config_path = 'config.' .. (config_relative_path)
-  local config_is_available, config = pcall(require, config_path)
-  if not config_is_available then
-    config = {}
-    if LOG_INTO.notify then
-      notify(
-        'Config for plugin `' .. plugin_name .. '` on path `' .. config_path .. '` does not exist!\nReverting to default configuration.',
-        { title = 'Core' }
-      )
-    end
-
-    log('Core')(
-      'Config for plugin `' .. plugin_name .. '` on path `' .. config_path .. '` does not exist!\nReverting to default configuration.'
-    )
-  end
-
-  local absolute_path = require('constants.env').NVIM_LUA_CONFIG ..'/'.. string.gsub(config_relative_path, '%.', '/')
-  if not require('utils.file').is_lua_module(absolute_path) then
-    plugin.setup(config)
-
-    return plugin_is_available, plugin
-  end
-
-  plugin.setup(config.settings)
-  for mode, keymappings_for_mode in pairs(config.keymappings or {}) do
-    require('config.Ui.which_key.utils').apply_keymappings_once_ready(mode, keymappings_for_mode)
-  end
-
-  return plugin_is_available, plugin
-end
-
-
 --- Set theme with rollback mechanism: colorscheme -> backup  -> fallback.
 ---@param colorscheme_name (string) Preferred colorscheme.
 ---@param backup_colorscheme_name (string) Custom theme from stable plugin.
@@ -124,7 +45,11 @@ local function load_coloscheme(colorscheme_name, backup_colorscheme_name, fallba
 end
 
 return {
-  prequire_plugin = prequire_plugin,
-  simple_plugin_setup = simple_plugin_setup,
+  prequire_plugin = require('ds_omega.utils.prequire_plugin'),
+
+  get_plugin_config = require('ds_omega.utils.get_plugin_config'),
+  apply_plugin_keymappings = require('ds_omega.utils.apply_plugin_keymappings'),
+  simple_plugin_setup = require('ds_omega.utils.simple_plugin_setup'),
+
   load_coloscheme = load_coloscheme,
 }
